requirements.txt
streamlit
google-generativeai
pandas
python-dotenv
PyPDF2
gTTS
openpyxl
requests
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
chromadb
tiktoken
